requirements.txt
aiogram==3.20.0.post0
matplotlib==3.10.3
python-dotenv==1.1.1
